remove S6 settings which are no longer necessary / startup service improvements (#168)
Some S6 settings for this container are changed from the default, this
is no longer necessary, so remove them. This means the container will
exit on its own instead of running into the usual 10 second timeout of
docker stop / compose down / compose up.

Simple fix for the timezone environment variable warning. (color code
printed instead of colored warning)

startup service:
- don't print the whole path via s6wrap
- start after legacy services (not necessary for this container, but
useful if this is ever moved to the baseimage)
- run scripts in /etc/s6-overlay/finish.d as down action (not necessary
for this container but useful for reuse)