allow persistent directory for tar1090-update
when recreating the container via docker compose, changes in the docker
image are lost. this means tar1090-update needs to download data again.
especially when changes to container are made via automation, this can
happen relatively often and it's desirable that the data are not
required to be downloaded more than once

change the directory for the tar1090 and tar1090-db git dirs and the
aircraft.csv.gz file to TAR1090_UPDATE_DIR=/var/globe_history/tar1090-update

this directory is already mounted on most installs and redownloads are avoided

the Dockerfile places aircraft.csv.gz in TAR1090_INSTALL_DIR so it can be
used in the unlikely event of /var/globe_history being mounted while
UPDATE_TAR1090 is false